requirements.txt
Django==4.1.6
matplotlib==3.6.3
pandas==1.5.3
seaborn==0.13.2
django-multiselectfield==0.1.12
django-filter==23.5
reportlab==4.1.0
django-auto-logout==0.5.1
django-cryptography==1.1
django-admin-logs==1.2.0
numpy==1.26.4